move function to liesel.model.goose
